+++ cmd/zetae2e/local/local.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
+	"regexp"
 	"syscall"
 	"time"
 
@@ -49,6 +50,7 @@ const (
 	flagSkipTrackerCheck  = "skip-tracker-check"
 	flagSkipPrecompiles   = "skip-precompiles"
 	flagUpgradeContracts  = "upgrade-contracts"
+	flagTestFilter        = "test-filter"
 )
 
 var (
@@ -87,6 +89,7 @@ func NewLocalCmd() *cobra.Command {
 	cmd.Flags().Bool(flagSkipPrecompiles, true, "set to true to skip stateful precompiled contracts test")
 	cmd.Flags().
 		Bool(flagUpgradeContracts, false, "set to true to upgrade Gateways and ERC20Custody contracts during setup for ZEVM and EVM")
+	cmd.Flags().String(flagTestFilter, "", "regexp filter to limit which test to run")
 
 	cmd.AddCommand(NewGetZetaclientBootstrap())
 
@@ -119,7 +122,10 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 		skipPrecompiles   = must(cmd.Flags().GetBool(flagSkipPrecompiles))
 		upgradeContracts  = must(cmd.Flags().GetBool(flagUpgradeContracts))
 		setupSolana       = testSolana || testPerformance
+		testFilterStr     = must(cmd.Flags().GetString(flagTestFilter))
 	)
+
+	testFilter := regexp.MustCompile(testFilterStr)
 
 	logger := runner.NewLogger(verbose, color.FgWhite, "setup")
 
@@ -191,6 +197,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 		conf.DefaultAccount,
 		logger,
 		runner.WithZetaTxServer(zetaTxServer),
+		runner.WithTestFilter(testFilter),
 	)
 	noError(err)
 
@@ -520,6 +527,14 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 		logger.Print("❌ e2e tests failed after %s", time.Since(testStartTime).String())
 		os.Exit(1)
 	}
+
+	// Default ballot maturity is set to 30 blocks.
+	// We can wait for 31 blocks to ensure that all ballots created during the test are matured, as emission rewards may be slashed for some of the observers based on their vote.
+	// This seems to be a problem only in performance tests where we are creating a lot of ballots in a short time. We do not need to slow down regular tests for this check as we expect all observers to vote correctly.
+	if testPerformance {
+		deployerRunner.WaitForBlocks(31)
+	}
+
 	noError(deployerRunner.WithdrawEmissions())
 
 	// if all tests pass, cancel txs priority monitoring and check if tx priority is not correct in some blocks