+++ e2e/e2etests/e2etests.go
@@ -178,10 +178,15 @@ const (
 	 Operational tests
 	 Not used to test functionalities but do various interactions with the netwoks
 	*/
-	TestDeploy                         = "deploy"
-	TestOperationAddLiquidityETHName   = "add_liquidity_eth"
-	TestOperationAddLiquidityERC20Name = "add_liquidity_erc20"
-	TestOperationAddLiquiditySPLName   = "add_liquidity_spl"
+	TestDeploy                                    = "deploy"
+	TestOperationAddLiquidityETHName              = "add_liquidity_eth"
+	TestOperationAddLiquidityERC20Name            = "add_liquidity_erc20"
+	TestOperationAddLiquidityBTCName              = "add_liquidity_btc"
+	TestOperationAddLiquiditySOLName              = "add_liquidity_sol"
+	TestOperationAddLiquiditySPLName              = "add_liquidity_spl"
+	TestOperationAddLiquiditySUIName              = "add_liquidity_sui"
+	TestOperationAddLiquiditySuiFungibleTokenName = "add_liquidity_sui_fungible_token" // #nosec G101: Potential hardcoded credentials (gosec), not a credential
+	TestOperationAddLiquidityTONName              = "add_liquidity_ton"
 
 	/*
 	 Stateful precompiled contracts tests
@@ -1219,6 +1224,24 @@ var AllE2ETests = []runner.E2ETest{
 		TestOperationAddLiquidityERC20,
 	),
 	runner.NewE2ETest(
+		TestOperationAddLiquidityBTCName,
+		"add liquidity to the ZETA/BTC pool",
+		[]runner.ArgDefinition{
+			{Description: "amountZETA", DefaultValue: "50000000000000000000"},
+			{Description: "amountBTC", DefaultValue: "5000000000"},
+		},
+		TestOperationAddLiquidityBTC,
+	),
+	runner.NewE2ETest(
+		TestOperationAddLiquiditySOLName,
+		"add liquidity to the ZETA/SOL pool",
+		[]runner.ArgDefinition{
+			{Description: "amountZETA", DefaultValue: "50000000000000000000"},
+			{Description: "amountSOL", DefaultValue: "50000000000"},
+		},
+		TestOperationAddLiquiditySOL,
+	),
+	runner.NewE2ETest(
 		TestOperationAddLiquiditySPLName,
 		"add liquidity to the ZETA/SPL pool",
 		[]runner.ArgDefinition{
@@ -1226,6 +1249,33 @@ var AllE2ETests = []runner.E2ETest{
 			{Description: "amountSPL", DefaultValue: "50000000000000000000"},
 		},
 		TestOperationAddLiquiditySPL,
+	),
+	runner.NewE2ETest(
+		TestOperationAddLiquiditySUIName,
+		"add liquidity to the ZETA/SUI pool",
+		[]runner.ArgDefinition{
+			{Description: "amountZETA", DefaultValue: "50000000000000000000"},
+			{Description: "amountSUI", DefaultValue: "50000000000"},
+		},
+		TestOperationAddLiquiditySUI,
+	),
+	runner.NewE2ETest(
+		TestOperationAddLiquiditySuiFungibleTokenName,
+		"add liquidity to the ZETA/SuiFungibleToken pool",
+		[]runner.ArgDefinition{
+			{Description: "amountZETA", DefaultValue: "50000000000000000000"},
+			{Description: "amountSuiFungibleToken", DefaultValue: "50000000"},
+		},
+		TestOperationAddLiquiditySuiFungibleToken,
+	),
+	runner.NewE2ETest(
+		TestOperationAddLiquidityTONName,
+		"add liquidity to the ZETA/TON pool",
+		[]runner.ArgDefinition{
+			{Description: "amountZETA", DefaultValue: "50000000000000000000"},
+			{Description: "amountTON", DefaultValue: "50000000000"},
+		},
+		TestOperationAddLiquidityTON,
 	),
 	/*
 	 Stateful precompiled contracts tests