+++ zetaclient/chains/evm/signer/outbound_data_test.go
@@ -47,7 +47,7 @@ func TestNewOutboundData(t *testing.T) {
 		assert.NotEmpty(t, out.height)
 		assert.NotEmpty(t, out.gas)
 		assert.True(t, out.gas.isLegacy())
-		assert.Equal(t, uint64(minGasLimit), out.gas.Limit)
+		assert.Equal(t, cctx.GetCurrentOutboundParam().CallOptions.GasLimit, out.gas.Limit)
 
 		assert.Empty(t, out.message)
 		assert.NotEmpty(t, out.cctxIndex)