+++ zetaclient/chains/sui/signer/signer_tx.go
@@ -63,7 +63,7 @@ func (s *Signer) buildWithdrawal(ctx context.Context, cctx *cctypes.CrossChainTx
 		Module:          s.gateway.Module(),
 		Function:        funcWithdraw,
 		TypeArguments:   []any{coinType},
-		Arguments:       []any{s.gateway.ObjectID(), amount, nonce, recipient, withdrawCapID},
+		Arguments:       []any{s.gateway.ObjectID(), amount, nonce, recipient, gasBudget, withdrawCapID},
 		GasBudget:       gasBudget,
 	}
 
@@ -78,9 +78,8 @@ func (s *Signer) broadcast(ctx context.Context, tx models.TxnMetaData, sig [65]b
 	}
 
 	req := models.SuiExecuteTransactionBlockRequest{
-		TxBytes:     tx.TxBytes,
-		Signature:   []string{sigBase64},
-		RequestType: "WaitForLocalExecution",
+		TxBytes:   tx.TxBytes,
+		Signature: []string{sigBase64},
 	}
 
 	res, err := s.client.SuiExecuteTransactionBlock(ctx, req)