+++ changelog.md
@@ -1,4 +1,16 @@
+# UNRELEASED
+
+### Features
+
+* [3672](https://github.com/zeta-chain/node/pull/3672) - zetaclient: cache tss signatures for performance.
+
 # CHANGELOG
+
+## Unreleased
+
+### Features
+
+* [3671](https://github.com/zeta-chain/node/pull/3671) - use gas budget argument to refund TSS for Sui withdraw cost
 
 ## v29.0.0